[dagster-airlift] Split out task operator fxnality
