+++ examples/experimental/dagster-airlift/dagster_airlift/in_airflow/__init__.py
@@ -1,5 +1,3 @@
-from .base_proxy_operator import (
-    BaseProxyToDagsterOperator as BaseProxyToDagsterOperator,
-    DefaultProxyToDagsterOperator as DefaultProxyToDagsterOperator,
-)
+from .base_asset_operator import BaseDagsterAssetsOperator as BaseDagsterAssetsOperator
 from .proxying_fn import proxying_to_dagster as proxying_to_dagster
+from .task_proxy_operator import BaseProxyTaskToDagsterOperator as BaseProxyTaskToDagsterOperator
