+++ examples/experimental/dagster-airlift/examples/tutorial-example/tutorial_example/custom_operator_examples/custom_proxy.py
@@ -3,11 +3,11 @@ from pathlib import Path
 import requests
 from airflow import DAG
 from airflow.utils.context import Context
-from dagster_airlift.in_airflow import BaseProxyToDagsterOperator, proxying_to_dagster
+from dagster_airlift.in_airflow import BaseDagsterAssetsOperator, proxying_to_dagster
 from dagster_airlift.in_airflow.proxied_state import load_proxied_state_from_yaml
 
 
-class CustomProxyToDagsterOperator(BaseProxyToDagsterOperator):
+class CustomProxyToDagsterOperator(BaseDagsterAssetsOperator):
     def get_dagster_session(self, context: Context) -> requests.Session:
         if "var" not in context:
             raise ValueError("No variables found in context")
